+++ Magick.NET.Samples/Samples/Magick.NET/ReadImage.cs
@@ -127,7 +127,7 @@ namespace RootNamespace.Samples.MagickNET
 
       // Read pdf with custom density.
       MagickReadSettings settings = new MagickReadSettings();
-      settings.Density = new PointD(144, 144);
+      settings.Density = new Density(144);
 
       using (MagickImageCollection collection = new MagickImageCollection(SampleFiles.SnakewarePdf, settings))
       {