+++ Changes.md
@@ -2,6 +2,7 @@
 
 ### Magick.NET 7.0.0.0104:
  - Removed Transform methods from MagickImage (use Resize and Crop instead).
+ - The Density property of MagickSettings now also uses the Density class.
 
 ### Magick.NET 7.0.0.0102:
  - Removed Resolution properties from MagickImage and replaced this with a new Density property.